chore: ✨ rename files to follow `x.server.ts` convention for Next.js
