+++ AGENTS.md
@@ -19,6 +19,7 @@ import { Activity } from "react";
 ```
 
 - Always add breadcrumbs using `@breadcrumb` in the page component (Parallel Routes).
+- Name `x.server.ts` files if inside is 'use server' code in Next.js.
 
 ### Improved Caching APIs
