+++ apps/docs/content/docs/dev/captcha/custom-adapter.mdx
@@ -90,7 +90,7 @@ export const FormSignUp = ({
 
 ### Submit form with captcha
 
-```tsx title="plugins/{plugin_name}/src/components/form/sign-up/mutation-api.ts"
+```tsx title="plugins/{plugin_name}/src/components/form/sign-up/mutation-api.server.ts"
 "use server";
 
 import type { z } from "zod";
