+++ apps/docs/content/docs/dev/captcha/index.mdx
@@ -151,7 +151,7 @@ export const FormSignUp = ({
 
 Next, you need to set `captchaToken` in your mutation API call. This token is provided by the `AutoForm` component when the form is submitted.
 
-```tsx title="plugins/{plugin_name}/src/components/form/sign-up/mutation-api.ts"
+```tsx title="plugins/{plugin_name}/src/components/form/sign-up/mutation-api.server.ts"
 "use server";
 
 import type { z } from "zod";
