+++ packages/vitnode/src/views/admin/layouts/sidebar/sign-out.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { DropdownMenuItem } from "@/components/ui/dropdown-menu";
-import { logOutMutationApi } from "@/views/layouts/theme/header/user/auth/log-out-mutation-api";
+import { logOutMutationApi } from "@/views/layouts/theme/header/user/auth/log-out-mutation-api.server";
 
 export const SignOutUserFooterSidebarAdmin = ({
   children,
