+++ packages/vitnode/src/views/admin/views/core/dashboard/send-notification/send-notification.tsx
@@ -7,7 +7,7 @@ import { toast } from "sonner";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 
-import { sendNotificationMutation } from "./mutation-api";
+import { sendNotificationMutation } from "./mutation-api.server";
 
 export const SendNotificationAction = ({
   defaultUserId,
