+++ packages/vitnode/src/views/auth/password-reset/form/use-form.ts
@@ -5,7 +5,7 @@ import z from "zod";
 
 import type { AutoFormOnSubmit } from "@/components/form/auto-form";
 
-import { mutationApi } from "./mutation-api";
+import { mutationApi } from "./mutation-api.server";
 
 export const useForm = () => {
   const t = useTranslations("core.auth.sign_up");
