+++ packages/vitnode/src/views/auth/sign-in/form/use-form.ts
@@ -5,7 +5,7 @@ import { z } from "zod";
 
 import type { AutoFormOnSubmit } from "@/components/form/auto-form";
 
-import { mutationApi } from "./mutation-api";
+import { mutationApi } from "./mutation-api.server";
 
 export const useFormSignIn = ({ isAdmin = false }: { isAdmin?: boolean }) => {
   const [error, setError] = React.useState<"" | "access_denied">("");
